接入Riptide Network

+++ Core/Network/PackId.cs
@@ -1,6 +1,6 @@
 ﻿namespace WeCraft.Core
 {
-    public enum PackId:uint
+    public enum PackId:ushort
     {
         C2S_Ping=1,
         S2C_Pong=2,
