+++ Server/StartUp/Setting.cs
@@ -3,7 +3,7 @@
     public class Setting:ISetting
     {
         public string Ip { get; } = "127.0.0.1";
-        public int Port { get; } = 25575;
-        public int TickRate { get; } = 20;
+        public ushort Port { get; } = 25575;
+        public byte TickRate { get; } = 20;
     }
 }
